chore: move utils to services
